requirements.txt
Django==4.0.4
django-redis==5.2.0
djangorestframework==3.13.1
hiredis==2.0.0
redis==4.2.2